compress descriptions with DTE
Digram tree encoding (DTE), also called byte pair encoding (BPE),
is a simple text codec.  It's not as powerful as Huffword, but
it's byte oriented (not bit oriented) so it works fast.
<https://en.wikipedia.org/wiki/Byte_pair_encoding>

Many English translation hacks of 16-bit JRPGs compress their
text with DTE, as do my NES port of robotfindskitten and my NES
and GB ports of 240p Test Suite.  The DTE implementation used
here is borrowed from RFK and 240p.

The collection that tests DTE is "dtetest", which has a full
deck of about 60 activities.