Donut speedup (#12)
For the following commits, JRoatch tried to reduce the total time
that `load_titledir_chr_rom` takes to load the CHR data of 'Miedow'.
Donut packs better than PB53 but is quite a bit slower, at 2.14
seconds.  Turns out roughly 0.42 s was shaved off, at the cost of
10 bytes of ROM and very slightly worse compression rate.

- Inline byte writing subroutines: -0.14 s
- Fast path for whole planes of $00 or $FF: -0.08 s
- Hoist computation out of loop by splitting it into 3 or more
  (0x00, 0xFF, PB8, rotated PB8): -0.18 s

The feature to mix decompression to an existing block is disabled.
It was intended for interbank redundancy in 'The Alfonzo Game', but
that'll have to wait for proper bank interleaving.  Disabling it
costs 64 bytes in Action 53 volume 4.

Potential savings for bypassing interbank_fetch later: 0.25 s

Also properly disable interrupts before starting the game, which was
a regression when CHR loading started calling `pently_update`.