Updated compressor for Donut and improved screenshot format
Older versions of Donut assumed that StopIteration exceptions can bubble up from generators. Which is a violation of PEP 479.
Along with that I tried an experiment with bitmaps of "don't care" bits where the compressor will fill in with bits that result in the best compression.

That experement was successful and saves 10% in screenshot data.