Freed up space in main menu bank.
quadpcm has a block size of 256 bytes, with a 1 byte header and 255 bytes of 510 samples. The original audio track was divisible by 512 sameples. This discrepancy cause a extra block with mostly padding to be genreated. So the audio was speed up slightly to fit within 26 blocks.

Additionally The KEYBLOCK, being practically the header of the game database, has been move to the beginning of the bank.

With the new space now available the RODATA segment is now moved back to menu bank.