Relocated selnow.qdp, and build now attempt to use faster Donut encoding
selnow.qdp is placed within the A53 database now that quadpcm can play fromanywhere.
It's at a fixed location for now, but it could eventualy be part of the build script.

If the fast C version of donut is in the system path, the build script will use it.
As a consequence the unutilized interleaving has been removed.
In the future interleaving may be reimplmented by some sort of LZ process with mabye a [backref,length] command in the donut format.