exitpatch: use anywhere patch everywhere if at all
If PC is in a bank that was patched by finish(), the $FFF0 patch
(or another patch ending with a jump at $FFF7) gets used.  But if
a BNROM/AOROM game's main bank was patched with the anywhere or
a53_anywhere patch, the NROM patch in the other banks will switch
to the last bank of the ROM, but the jump at $FFF7 won't be there.
This means the other banks of the same ROM need to use the anywhere
patch, which puts this jump in RAM.

Issue was reported by jroatch via Discord, affecting the BNROM game
Super Homebrew War.